Add login flag to container image prepare defaults
If a registry requires authentication and --local-push-destination is
not used, a user may need to enable the login functionality. This change
adds a --enable-registry-login flag to `openstack container image
prepare default` that can be used to ensure the
ContainerImageRegistryLogin flag is defined as true.  Previously this
would have needed to be added in elsewhere.

Change-Id: I809023604fe119ba162638d921ffec56d4eb72f2