Initially set rc to 0, not 1
With rc being set to 1 initially, it was never reset to 0 when using
--output-only. This meant when running tripleo deploy with
--output-only, it always exited 1, causing an exception.

This patch instead initially sets the value to 0. The only other place
where rc is used is to capture the return from running ansible, so if
that returns non-zero, rc will be updated accordingly so that the
deployment will fail.

blueprint tripleo-deploy-multinode
Change-Id: Ie116b4c600317d618a9e55b91a945d03e71fb780
Signed-off-by: James Slagle <jslagle@redhat.com>