Detect TLS for pre-provisioned nodes
In cases where we need to support TLS for pre-provisioned nodes, we need
to be able to enroll hosts without using novajoin. This commit detects
if an OTP is present but without novajoin enabled. If so, if uses a
separate composible service, without novajoin, to enroll the undercloud
as a FreeIPA host.

Change-Id: Id6d193ab10e43a5d2706705588269654073910cc