Don't set DeployIdentifier in tripleoclient
We set this to a new unique value or empty string
based on skip_deploy_identifier in tripleo-common.
As we don't have a way to compare an old DeployIdentifer
with a new DeployIdentifier in ansible, there is no
point in using the existing value from stack.

Change-Id: Ic52ce8ff38394df9f5234c257ce1affee7ff0791
(cherry picked from commit e001960d1588b64afc939f5b3aea822b913eef92)