Add "yes" prompt for update/upgrades commands
Update/Upgrade commands have now a prompt by default that ask for
confirmation before proceeding. It'll prevent an user to run the
command that may cause the problems to infrastructure.
This prompt can be skipped with --yes/-y argument.

Note: putting "UPDATE" and "UPGRADE" in uppercase to make sure this is
visible and clear. We have seen many users running the wrong command and
ending up doing an upgrade instead of an update.

Note2: this prompt will be ported to the upgrade and FFWD workflows to
prevent unexpected execution to prevent potential harm to
infrastructures.

Depends-On: https://review.opendev.org/741657
Change-Id: I838e6748879c668dd004ca2243b7b00b857c2a7b
(cherry picked from commit 7a044ca2517d26ea4ea763cc76d373ad4b39f7f9)