Refactor upgrade prepare and converge.
As we aren't required to re-deploy any more during the converge step
the actions performed by the Upgrade Converge and Prepare are exactly
the same, the only difference is the lifecycle environment file we
use to set and unset parameters.

This patch refactors both classes, the UpgradePrepare and UpgradeConverge
clasess:

- UpgradePrepare: Stop forcing update-plan-only and then try to do most of the
actions from deployment into UpgradePrepare. This change will force config_download
to false (which will avoid running the full deployment and do only an stack update),
then download the config by invoking get_config from utils and lastly enable ssh
into overcloud nodes, as that part is perforemd in deployment only if config_download
is set to true. Add new attributes to UpgradePrepare so we can override them in
UpgradeConverge via inheritence.

- UpgradeConverge: Inherit from UpgradePrepare instead of DeployOvercloud. Set the
right value for the class attributes and let UpgradePrepare class do all the magic.

Change-Id: I6148511eb6ad1e3798a7bf40c721824830c7073d