overcloud_deploy: raise at the end if Ansible Runner had exception
In the context of MaxFailPercentage, if a node (e.g. one of 100
computes) has any fatal error during an Ansible play and we tolerate
some percentage of failure; we want to raise the error but at the very
end of the deployment.

So this patch puts the following actions in a "finally" block so they
always execute:

* Create overcloudrc
* Sets the right permissions to the clouds.yaml
* Execute _deploy_postconfig
* Print infos like Keystone entpoint, Horizon URL, RC file location
  and deploy message (saying if there is an error or not)

And at the very end we raise the actual AnsibleRunner trace if the
deployment failed.

So even if a node failed and we tolerate it, we'll still finish the
deployment until the end, but for UX purpose we want to raise it at the
very end.
Note that when it fails and it's tolerated, Ansible prints the node as
"ignored":

PLAY RECAP ******************************************************************
compute-0    : ok=555  (...)  failed=0    skipped=484  rescued=0    ignored=0
compute-1    : ok=60   (...)  failed=1    skipped=40   rescued=0    ignored=1
controller-0 : ok=960  (...)  failed=0    skipped=709  rescued=0    ignored=0
controller-1 : ok=920  (...)  failed=0    skipped=693  rescued=0    ignored=0
controller-2 : ok=919  (...)  failed=0    skipped=693  rescued=0    ignored=0
undercloud   : ok=86   (...)  failed=0    skipped=57   rescued=0    ignored=0

To improve UX, we'll investigate an Ansible callback to properly tell
what nodes needs to be re-deployed.

Note: also mock copy_clouds_yaml since it wasn't tested before but it's
failing to reach the files on the filesystem.
Change-Id: I7d733499e74abe2cdf91526df608dc7c273bf19e