Fall back to local upload when there is no glance endpoint
When there is no glance endpoint, it is currently necessary to use
the --local argument when running "openstack overcloud image upload"
to copy the files to /var/lib/ironic/images.

With this change, if there is no glance endpoint then the upload
command will run as if the --local argument is specified.

Change-Id: I264c26f8a63f36cb921c06d1d01db55383394737
Blueprint: nova-less-deploy