[TRAIN ONLY] Do not encode hostnames in validation result
Encoding the names to utf-8 leads to unwanted failures such as:
KeyError: b'undercloud'
as well as
TypeError: sequence item 0: expected str instance, bytes found

Removing the two calls to .encode() solves those issues.

It also correct a tiny deprecation warning, moving from LOG.warn to
LOG.warning.

This is a train-only patch since later the whole thing was moved out of
tripleoclient in favor of validations-libs and/or validations-common.

Change-Id: Ie1530321b488bef11692553ea25131d9360763dd