Fix package manager used during undercloud packages update
As described in the related bug we need to use dnf instead of yum
for centos8 otherwise it fails in the undercloud packages update.

I am using the python version to determine - py3 is dnf. As part of
discussion here when dnf isn't available fall back to yum.

Found as part of the work in [1].

Related-Bug: 1886837
[1] https://tree.taiga.io/project/tripleo-ci-board/task/1817

Change-Id: Idac62d37a19ee49f30936e20021a9dab5af40eec
(cherry picked from commit 9c53cb3ef5ad513b05d4888552a3280e0eec43ca)