Cleanup UndercloudHostsEntries
When we update the UndercloudHostsEnries we basically
take the output of 'getent hosts "$(hostname -s).ctlplane"'
and push it into a parameter, so tripleo-ansible can
make sure it adds the hostentry for the undercloud on
the whole overcloud.

The problem is that getent hosts can return multiple
entries, which will then be injected into the parameter
and then written into /etc/hosts. This contstantly adds the
undercloud.ctlplane string and we end up adding it at
everydeploy making it grow quadratically.

This eventually leads to a too large json file and the deploys
start failing with:
    heat.common.exception.RequestLimitExceeded:
      Request limit exceeded: JSON body size (4396634 bytes)
        exceeds maximum allowed size (4194304 bytes).

Tested this by deploying an environment and then running
a few redeploys to make sure that the undercloud entries
in /etc/hosts on the undercloud itself do not grow at
each redeploy.

Change-Id: I37d75600825f48be9e15470cacba7b3a0371a3e2
Closes-Bug: #1887165
(cherry picked from commit 2873dd4df62f104bef6a9ce36bdf5385b75b96b5)