Correct ansible playbook execution handler
This change modifies the `run_ansible_playbook` method so that its correctly handling
output. When a playbook has a lot output the method will hang in an S+ state waiting
for stdout to be read and the FD closed. To correct this issue the method has been
modified to return only the rc information and allow the `run_command_and_log` to
handle stdout, which ensures this subprocess invocation is handled in the same way
as all other invocations.

> All uses of `run_ansible_playbook` have been updated to ensure we're handling
  the process correctly throughout the client.
> The method `run_ansible_playbook` will now execute and return None. In the event
  of a failure, the method will still raise a "Runtime" exception and log a warning.
> This change only impacts stable/train, this issue has been resolved in future
  releases through the use of Ansible-Runner.
> Tests have been updated to support the changes within the `run_ansible_playbook`
  method.

Change-Id: Id60bf618f6cd778970c5ff3aee19bcc40d294c3b
Signed-off-by: Kevin Carter <kecarter@redhat.com>