[TRAIN-Only] Increase deploy workflow wait
During the execution of the deploy plan, we will also run container
image prepare in a dry run. This may still reach out to determine tags
which can contribute to the process taking longer than 10 minutes in CI.

We are doubling the wait time from 10 minutes to 20 minutes for now.
Eventually we should probably not exit if the task is still running.

We dropped this code in Ussuri so it only applies in Train.

Change-Id: Ifa94bffcf4fadc5d4a87675b5d4a3357e8808ac3
Parial-Bug: #1891071