[train-only] Add ipa config check to _cleanup_ipa
This adds a check to determine if this system is a properly configured
ipa client, if not the _cleanup_ipa code will be bypassed. This should
eliminate the need to pass -s on non TLS-e environments.

This commit also defensively bypasses IPA cleanup if tripleoclient can't
generate a reasonable inventory file. It's better to bail early and
inform the user that they need to manually clean-up IPA instead of
munging the data in IPA based on assumptions. This is especially true if
IPA is a shared resource and not specific to the OpenStack deployment.

Change-Id: I093faee1d136d1fa697d45647c8311fe82379c5c
BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1865900