TCIB: add --rhel-modules argument
--rhel-modules is a new argument that will let us override the hardcoded
RHEL module versions that are provided in the base image.

For example, if we want to build the images against a specific module
version, we'll now be able to do the following:

  $ openstack tripleo container image build --rhel-modules \
        redis:5,virt:8.3

Note: for now we're opinionated about the RHEL modules that we support
and the --rhel-modules won't allow to install any modules, but only the
ones that we support: ['container-tools', 'mariadb', 'redis', 'virt'].
The reason is that in theory all necessary modules should be provided in
the base image layout in tripleo-common, and the use case here is being
able to change their version from the CLI, specially for our release
challenges with future OSP zstreams.


Change-Id: Icaeabb2e7cb7adbbd7d2ef563dd613f2aacf8fe3
(cherry picked from commit a3b3aeb93b16bea51a53083e0b53faf5c15ee2bb)