Allow the container build tools to run config-less
This change allows the container build tools to run without a
provided configuration file. The container configuration file
still in service was derived from our old kolla implementation
which no longer needs to exist. The build process will now
simply walk the build path and build all containers found when
a configuration file is not defined or present. Logging has
been added to inform the user of the tools discoveries which
should provide a nice user experience when running container
builds.

In the event a container configuration file is used, the tools
will now ensure all dependent images are inherently part of the
build preparation.

Tests have been updated to ensure that the client is operational
in both a config and config-less situation.

Change-Id: I89aebb0ae9734b0371bd3fdbd34643babfcca8f3
Signed-off-by: Kevin Carter <kecarter@redhat.com>