Add --temp-dir to overcloud image build
By default, DIB puts the temp folders it uses in /tmp. In newer
versions, this can cause images to go missing because systemd will
cleanup tmp (see tmpfiles.d) which may cause files and directories in
the image to missing. This change adds a --temp-dir option to the cli
which by default will configure TMPDIR to be the current working
directory if not specified.

Change-Id: Ib5f82a175266675d9639416dd664604bee370a1d
Closes-Bug: #1879766