Remove c7 containers multinode job
This job was kept initially to maintain coverage on c7, since
there is no user of CentOS-7 train, so we are removing this
job.

A periodic version of the job will be still running in the
upstream infra, https://review.opendev.org/#/c/752029/

Removal of this job is planned as a part of moving to
centos-8 based content provider jobs.

Change-Id: I6a741a2f61faf048763aafe0a246bfd0d153aabc
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>