Fix tcib build
The tcib build process assumes that all the container configs exist in a
'tcib' folder as defined via --config-path. This change documents this
requirement and additionally we've hard coded 'container-images' in
the image config to exclude this folder however if you define a
custom --config-path, this fails.  This change removes the hard coded
'container-images' while documenting the 'tcib' requirement.

Change-Id: Iad987da5bedc30da7118a3aec5dbe96cd8d71d1e
Closes-Bug: #1897616