Generate routes for undercloud ctlplane network attrs
In https://review.opendev.org/753195 we set up ctlplane
network attributes and later use those in THT when setting
group_vars for os_net_config templates in ansible.

The change missed to add 'host_routes' for peer-subnets in
a spine-and-leaf set-up. This caused introspection and
provisioning to fail in spine-and-leaf set-ups because the
undercloud did'nt know how to reach the remote subnets.

This change updates the code to include calculated routes
to the remote subnets.

Change-Id: I265b2b586ceaeaa98bbf6073bb79cde6a91627da
Closes-Bug: #1899008