Use default environment with --deployed-server
We still expect users to add -e deployed-server-environment.yaml
to the deploy command when using --deployed-server (ex. when
using metalsmith for node provisioning). This adds the default env
automatically when using --deployed-server with the deploy command.

The resource mappings and parameters in it can be overridden by
any environment file provided by the user.

Change-Id: I849e4077b7fba88651904642fa8c2df08ec8352a