[Validator Show Run] Display simplified results by default
This patch is displaying the content of the validation_output dictionary
which is a simplified output content (the former validation_output)
instead of dumping the entire log file in the stdout.

It also fixes the --full argument which was True forever!

Change-Id: Id94132d23d272d383b65a839a75b08f855fb6be3
Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
(cherry picked from commit 4229bee94016ad5a78dd8455f282a8e7f4a3d28e)