Allow playbooks to come from a venv version
Today if you pip install tripleoclient and pip install tripleo-ansible,
you can't run some of the cli commands because they cannot be found. In
order to support this, we need to check if the playbook folder is
available in a venv and fall back to the /usr/share/ansible version if
it does not exist.

Change-Id: I27ac01aae7155cfc84bf5aef926ae220b57373c9