Do not require authentication for the validator CLI
Since most of the actions of the validator is based on the filesystem,
we really don't need to require auth.
The only subcommand that might require it is the "run", since it might
go through the inventory generation.

Also, pass the "--os-cloud" parameter down to the
tripleo-ansible-inventory script. Value is hardcoded, since it is an
entry name we can't set (yet). It allows to authenticate only for the
relevant part of the validation run.

Resolves: rhbz#1892323
Closes-Bug: #1902052

Change-Id: I32c1e006d10ceb541aadb486befd8aa8fc6aa045