Ensure build packages installed at runtime
As an alternative to switching this process to ansible, we can handle
the installation of the required packages for the overcloud image
building process in the command itself. This change adds an install for
tripleo-common and the image element packages required to build the
overcloud images as part of the `openstack overcloud image build`
command itself.

Change-Id: I32eb04c81cba0f9688ddf6e40874790f033e84e8