Map build host repos to distro.repos.d instead of yum.repos.d
Currently, when building container images, we are overriding the ubi
image repos with the repos configured in the host. Effectively, this is
removing ubi repo config from the image build environment which most of
the times is not a problem as centos is on sync with ubi, but in some
cases as after RHEL 8.3 release, ubi is ahead of CentOS and need
packages from the ubi repo itself.

This patch is bind mounting the repos config from build host in
distro.repos.d instead of yum.repos.d. In that way, both the repos
configured by default in the image (ubi repos) and the ones in the host
are available.

Depends-On: https://review.opendev.org/c/openstack/tripleo-common/+/765077
Closes-Bug: #1902846
Change-Id: I55f4b8cf2ad971e196fbd06db743d07ce3135ea8
(cherry picked from commit a4fdf46d83a91e8cbe8f40abc79d94a1c31dc592)