Run validation without authentication in particular context
Since we are using mistral to get the list of the plan
in TripleO, we require authentication for the clients
otherwise the validation Run won't work due to unable
to get mistral context.
This patch handle the authentication if provided by the user and
environment, if no authentifcation, since we cant generate dynamic
inventory, it checks if a static is passed with the CLI, if not it
raised an exception.

If authentication is provided and succeed, then tripleoclient can
safely call the plan_management to check and ensure if the plan is
present.

Change-Id: I5a702a84ca0e1942f2e1636d31acc391f689a8ad