[TRAIN-only] Don't stop scale down if ansible fails
This is train only as the fix in future versions may be different.

If the ansible action fails, we still want to continue with the node
removal. Currently if all nodes being removed are down, the scale down
action which is invoked using the config download workflow will error.
There is no way to remove the nodes if we block on this.  Scale down
actions should be best effort so we should just throw a warning and
cleanup the node(s). In future versions we could add better handling of
specific error codes via ansible-runner, but the mistral execution is
not flexible enough to skip only one some errors.  We had addressed this
condition previously, however it appears that newer versions of ansible
are returning a non zero rc even if we have ignore_unreachable set to
true.

Change-Id: I1fe1fecffcf3b44895721118d675205e612155eb
Related-Bug: #1887702