[TRAIN-Only] Reduce open mock scope
There appears to be a race condition related to language parsing that is
being uncovered when we mock open around the arg parse code. This
highlights that we're relying to open mock instead of mocking a function
that uses open.  Rather than rewrite all the tests, let's reduce the
scope of the open mock to the action execution instead.  This only
really applies to train and older because we rewrote many of these
actions to ansible executions instead.

Change-Id: I7f7b25aa76bef692991200ba621a536f8868200a
Closes-Bug: #1909118