Add new option for `setup-only`
This change implements the ability to run all of the required setup
tasks to enable manual playbook execution. When the option is invoked,
the following workflows will be run:

* Generate the new Ansible content
* Downlad the Ansible content
* SSH key validation and application to overcloud nodes

This new option allows us to retain the current logic of stack-only
which will ensure legacy workflows are not interupted while also
enabling a more ansible centric user experience.

rhbz: 1873246
Change-Id: I3c0466aeed8cd7d2248c0d769b24821110a0e0a5
Signed-off-by: Kevin Carter <kecarter@redhat.com>