Write ansible-playbook-command.sh to workdir not playbook_dir
The reproducer command, ansible-playbook-command.sh, should not be
written to playbook_dir, as that directory may not always be writeable.
It is in fact commonly set to /usr/share/ansible/tripleo-playbooks,
which is not writeable by a non-privileged user.

Instead, write the command to workdir, which is either the
config-download directory, or the directory setup for use by ansible
runner.

Change-Id: I4ac0c2aaf8b6a8f4a9b5e5f40fa68929be32ee38
Signed-off-by: James Slagle <jslagle@redhat.com>