Add faultwrap filter to paste pipeline for heat-all
Without using the faultwrap filter, exceptions raised by heat-engine are
not translated into HTTPException subclasses that are understood by
webob.

This causes the exceptions to not be appropriately handled and the
correct HTTP codes are not returned to the client. Instead the exception
is assumed to be a 500 Internal Server Error instead of an informative
HTTP code (such as 404 for stack not found). Several spots in
tripleoclient/tripleo-common rely on known HTTP codes (such as 404)
being returned, so this commit adds the faultwrap filter to do the
translation.

Change-Id: I9cb25e6de5acce7e433aa559ca515ee7f3eb8dbf
Signed-off-by: James Slagle <jslagle@redhat.com>