Predictable dir for heat launcher
Adds an option to use the top level directory passed to the Heat
launcher class as the working directory for the ephemeral Heat process.
The default behavior remains, which is to create a tmp directory under
the top level specified dir.

The option to use a predictable directory will allow for (re)specifying
the same directory to restart the ephemeral Heat process with the same
configuration, logs, and db.

Change-Id: I16c53281e6fbfc497730bebb166acc21c5dbf947
Signed-off-by: James Slagle <jslagle@redhat.com>