Use a consistent working dir for deployment
Adds a new cli arg --working-dir, which defaults to:
$HOME/overcloud-deploy-<stack>

The working directory will be used for all files created by the
overcloud deploy command, instead of using tmp dirs and files directly
in $HOME. The working dir provides a single dir for all state associated
with the deployment, which is needed with the transition to using
ephemeral Heat, and especially when combined with
multi-stack/multi-overcloud.

This patch addresses:
- deployment status
- templates
- config-download
- heat-launcher
- overcloudrc

Further patches will address other uses of files outside of working-dir
and migrate them over.

Change-Id: I0d803f695c725c58ef2e6b655753b6c8248d1b2f
Signed-off-by: James Slagle <jslagle@redhat.com>