[Train] Bump queue subscription ttl for WebsocketClient
At present it's set 1t 10000s which is not enough
for many large deployments with current ansible
performance issues. Client gets stuck, where as
config-download workflow continues to success in
the background.

Though it can be made configurable, let's
bump it to 12 hrs (43200s) as it's only required
till we fix ansible performance issues.

Closes-Bug: #1872823
Change-Id: I02866d46cce0f6b8c7f58db121707e92f4559a8e