Add parameter to specify ca_cert
When generating system certificates, the templates set the values
of InternalTLSCAFile and PublicTLSCAFile to the file path for the
local CA cert created by certmonger.

This value is invalid for an IPA generated certs.  This patch sets
these values correctly to the default IPA CA cert location
(/etc/ipa/ca.crt)

Change-Id: Iad3a51bc853d3e868fd38dbeca8c20dd32725723