UX improvement when no nodes are found
When an operator executes a client command targetting the overcloud
the client performs a host lookup which can fail there are no hosts
returned from a given stack. This change implements a check that
the object has some value before attempting to iterate through the
nodes. If the object is null, an exception is raised to inform the
operator there was an error, and provides insight on how they might
resolve the issue. This is an improvement over the current setup
which simply results in an AttributeError should no nodes be
returned from a given stack.

Change-Id: Ib42a655a93a647651a20f1d6ac42fcb96bb4c271
Signed-off-by: Kevin Carter <kecarter@redhat.com>