Absolutize path for CIP output env
In the case where a relative path, or just a filename, is passed as
--output-env-file to container image prepare, the output file ends up
relative to the cli-container-image-prepare.yaml playbook.

In almost all cases, this is not what is expected, especially as the
output file could end up in the playbook directory. Instead, absolutize
the path, so that the output file ends up relative to the working
directory from where the command was run.

Signed-off-by: James Slagle <jslagle@redhat.com>
Change-Id: I9227365ef2900b1f077ac1abdcb4443957b15958