Restore old config-download symlink to new location
Change I5fc78158bbff468576eaddf1924e67a6666213e3 switched the symlink
source/target of the config-download dir for the consistent working dir
from new->old instead of old->new.

We want the old location at ~/config-download/<stack> to be a symlink to
the new location, and store the actual files in the new location. The
symlink is for backwards compatibility. A migration is added for any
files at the old location to the new location, if the new location does
not yet exist.

This patch also changes the default working dir to be a directory under
~/overcloud-deploy for the stack name instead of
~/overcloud-deploy-<stack>. Given multi-stack and multi-overcloud
support, defaulting to using a single directory is a more
straight forward approach and easier to collect.

Change-Id: I189bd751066cfaebd951d92d8371250aac5b4e9b
Signed-off-by: James Slagle <jslagle@redhat.com>