Use MultiKeyActionValue action for extra-[env-]vars arguments
This patch adds the ability to pass extra variables and extra
environment variables to the Ansible execution as
key1=value1,key2=value2 instead of using a dict.

This action ensure that ``dest`` is a list. The list will finally
contain multiple dicts, with key=value pairs in them.

This patch also enables the Ansible execution verbosity when using a
different callback by passing --extra-env-vars
ANSIBLE_STDOUT_CALLBACK=default in the command line.

Finally, this change will make the CLI easier to use and faster to type.

Co-Authored-By: Gaël Chamoulaud <gchamoul@redhat.com>
Co-Authored-By: David J. Peacock <dpeacock@redhat.com>

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: I2432db16f92b3aded51727393d6d8dbfbd61a257