Replace exception with LOG.info
This is a backport of [1] but in between now and then the code in
tripleo_validation was move back into python-tripleoclient[2] and the
review in [1] was included.

This means that we cannot have a review in master as it's already
there.

[1] https://review.opendev.org/c/openstack/tripleo-validations/+/777220
[2] https://review.opendev.org/c/openstack/python-tripleoclient/+/782776

Original commit message:

Updated _run_validator_run to no longer raise an execption if/when a group is empty

Using a log vs. exception is useful when we want to run an automated set of validations
within our Update CI pipeline.

Co-authored-by: Owen McGonagle <omcgonag@redhat.com>
Co-authored-by: Daniel Bengtsson <dbengt@redhat.com>
Change-Id: I1046148da87712934184373a10fa3d2a46e620e4
(cherry picked from commit 0debcde236a75a501eb47f11b31508a28d795271)