Check if both swift and rgw are enabled during Upgrade Prepare
As per [1] RGW is now deployed by default when the cephadm envs
are included. However, during the upgrade procedure we should
warn users if both services are enabled, failing early during
the update or upgrade prepare step and suggesting that the
cephadm-rbd-only env should be added to make sure Swift is kept
and RGW is not automatically included in the heat stack.
This change introduces this kind of check, and an exception is
raised if the stack already exists, Ceph is deployed by TripleO
and swift is found in the previosly created stack.

[1] https://review.opendev.org/786827

Change-Id: I32387f19346697655355e15b3cf4ff41bd303b08